Bug 15667: Fix date and branch formatting of messages in patron account display
This patch fixes the date and branch formatting of messages in patron
account display.

To verify:
Add a new message to a patron account from the 'checkouts' tab in staff,
the displayed date is not correctly formatted and the branch displays as code
instead as branch name

To test:
Apply patch, verify that the messages and branch names are properly formatted.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as described. The message dates in checkouts tab follow dateformat sypref

Signed-off-by: Kyle M Hall <kyle@bywatersolutions.com>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com