Bug 15784: Library deletion warning is incorrectly styled
When deleting a library in Administration ->  Libraries and groups, the
deletion confirmation message is not styled with the standard "dialog
alert" <div>. This patch modifies the confirmation markup to match the
standard.

Revised to include Font Awesome icons. Depends on Bug 15785.

To test, apply the patch and go to Administration -> Libraries and
groups.

- Try deleting a library. The confirmation box should have the standard
  "dialog alert" class.
- Test the 'confirm' and 'cancel' actions and make sure both complete
  correctly.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Buttons display as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com