Bug 15670 - Rename "Cancel" to "Cancel hold" when checking in a waiting item
When checking in an on-hold item which is waiting, the "Cancel" button
cancels the hold, so it should say so. This patch changes the text.

This patch also moves the 'cancel hold' button to the middle, on the
theory that buttons on the left should default to 'confirm' and buttons
on the right should generally imply 'cancel'--as in 'do nothing.' Moving
it to the middle seems like a good place to put it where users will have
to think about clicking it.

To test:

1. Find an hold which has been made 'waiting,' or check in an confirm a
   pending hold at your library.
2. Check in that item again and verify that the 'Hold found' dialog
   contains a 'Cancel hold' button which works to cancel the waiting
   hold.
3. Repeat the process and confirm that the functionality of the other
   buttons is unchanged.

Followed test plan, works as expected.
Signed-off-by: Marc Veron <veron@veron.ch>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com