Bug 15804: [Revised] Use standard dialog style for confirmation of MARC subfield deletion
When deleting a MARC framework subfield, the confirmation message is not
styled. This patch modifies the template to style the confirmation
message like similar ones.

Depends on Bug 15785 - Use Font Awesome icons in confirmation dialogs

To test, apply the patch and go to Administration -> MARC framworks.

 - Click 'MARC structure' for a framework you can edit.
 - Click 'subfields' for a tag you can edit.
 - Click 'Delete' for a subfield you can delete. You should see a
   confirmation dialog styled like other "alert" type dialogs.
 - Test both canceling and confirming deletion to confirm that both
   function correctly.

Revised to fix conflicts with rollback of Bug 13618

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as expected. Revised according with bug revert 13618

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com