Bug 15833: item.biblioitemnumber returns a biblioitem DBIx::Class object
2 links are wrong on the renew page, because item.biblioitemnumber does
not return the biblioitemnumber as expect the template.

Test plan:
Renew an item and check the item link.
It should contain a valid biblioitemnumber for the 'bi' parameter.

Signed-off-by: Owen Leonard <oleonard@myacpl.org>

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>
Link repaired. At first glance, I am wondering if biblioitemnumber is
of actual use here; is it only passed back and forth between script and
template?

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com