Bug 15828: Upload patron images is hard to read
The styling of the patron image upload form causes the text to be
smaller than it should be for the main body of a page. This patch
revises the form style and cleans up the markup a bit.

This patch also adds client-side validation of the form so that a file
upload is required, and a card number is required if an image file is
selected.

To test, apply the patch and go to Tools -> Upload patron images.

1. Confirm that the text in the form is the correct size.
2. With "Zip file" selected, confirm that submitting the form is blocked
   and the file upload marked as required.
3. With "Image file" selected, confirm that submitting the form
   with an empty card number field is blocked and the card number
   field is marked as required.
4. Confirm that uploading zip files and single images still works
   correctly.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com