Bug 15843: Move MARC subfields structure JavaScript into separate file
The JavaScript embedded in the MARC subfields structure template is not
dependent on template processing and can safely be moved to a separate
file. This patch does so. Removed in the process is some lines which
were commented out.

To test, apply the patch and go to Administration -> MARC framworks.

 - Click 'MARC structure' for a framework you can edit.
 - Click 'subfields' for a tag you can edit.
 - Click 'Edit' for a subfield you can edit.
 - Confirm that tabs and accordian sections work correctly.
 - Make changes under Advanced constraints -> Visibility and confirm
   that they are saved correctly.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com