Bug 15846: Move MARC Framework JavaScript into separate file
The JavaScript embedded in the MARC framework template is not
dependent on template processing to such an extent that it can't
be safely moved to a separate file. This patch does so, adding
definition of a few translatable strings to the template.

To test, apply the patch and go to Administration -> MARC frameworks.

- The table of existing frameworks should be sortable.
- The "Export" link should trigger the export modal.
- The "Import" link should trigger the import modal.
  -- Clicking "Import" without selecting a file should trigger an alert.
  -- Selecting the wrong file type should trigger an alert.
  -- Importing a file of the correct type should trigger an overwrite
     warning ("Do you really want to import...?").
  -- A progress indicator should show during the import, with correct
     values for "Importing from" and "Importing to".
  -- Importing an incorrectly-formatted file of a correct type should
     trigger an alert, "Error importing the framework..."

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised. Export and Import options work properly

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>