Bug 15671 - Show branch name instead of branch code in checkout history
When viewing the checkout history for a title the branchcode is
displayed. This patch alters the template to show the library name
instead.

To test, apply the patch and locate a title in the staff client which
has been checked out at least once. From the detail page, click
"Checkout history" in the sidebar. The "Checked out from" column of the
checkout history table should show the library name instead of the
branchcode.

Works as expected.
Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com