Bug 15867: Move MARC modification templates JavaScript into separate file
The JavaScript embedded in the MARC modification templates template is
not dependent on template processing to such an extent that it can't
be safely moved to a separate file. This patch does so, adding
definition of a some translatable strings to the template.

Other minor changes: A couple of JSHint warnings corrected, a couple of
missing semicolons added to template markup.

To test, apply the patch and go to Tools -> MARC modification templates

- Confirm that The add/update form works normally with various actions
  (Add/update,Delete,Move, etc.).
- Confirm that template switching works correctly.
- Confirm that template deletion confirmation works correctly.
- Confirm that cancelling an edit works correctly

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com