Bug 15916: Datatables - buttons area should be defined
Commit 243b797dd11bc0ac1f475195ab765ffbf2220f8b
    Bug 15285: Update common files

Set 'B' to the default dom of Datatable constructor.
This cannot work if there is no button defined.

This patch defined an empty 'buttons' param for this constructor, DT
does not raised an error anymore when initialising.

Test plan:
Confirm that the tables on the following tables are not broken:
- circ/waitingreserves.pl
- circ/reserveratios.pl
- patron_lists/lists.pl
- virtualshelves/shelves.pl
- acqui/invoices.pl
And with datatables:
- circ/circulation.pl
- admin/currency.pl

Signed-off-by: Marcel de Rooy <m.de.rooy@rijksmuseum.nl>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com