Bug 15950: Use Font Awesome icons for acquisitions basket close confirmation
When closing a basket in Acquisitions the confirmation dialog should be
styled using Font Awesome icons. This patch adds them.

The cancel button has been moved into its own form so that we can get
rid of several problematic JavaScript issues: The use of the
"javascript:" pseudo-protocol, the use of an "onclick" attribute, and
the use of a JS redirect where a GET action will do.

To test, apply the patch and locate an open basket in Acquisitions which
can be closed.

- Click the 'Close this basket' button and confirm that the icons in the
  dialog look correct.
- Test the cancel operation to make sure you are redirected back to the
  page showing the basket details.
- Test the confirm operation and verify that the basket is closed.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>
Please note that the confirmation message display depends on
BasketConfirmations.

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>