Bug 15960: Use Font Awesome icons for classification filing rule deletion error
When deletion of a classification filing rule is impossible, the dialog
saying so should be styled with Font Awesome icons. This patch corrects
this omission from the patch for Bug 15785.

Also changed: The dialog is now an "alert" style dialog instead of a
"message" one.

To test, apply the patch and go to Administration -> Classification
sources.

- Locate a classification filing rule which is in use by a
  classification source.
- Click the "Delete" link for that rule and verify that the resulting
  dialog is correctly styled.
- Click the "OK" button and confirm that you are correctly redirected.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertise

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com