Bug 15961 - Use Font Awesome icons for confirmation of currency deletion
When deleting a currency, the confirmation and error dialogs should use
Font Awesome icons. This patch makes this change.

To test, apply the patch and go to Administration -> Currencies and
exchange rates.

- Click the "delete" link next to any currency.
- Verify that the confirmation dialog is correctly styled.
  - Test the both cancel and confirm operations.
  - Verify that the deletion confirmation dialog is correctly styled.

Bug 15962 prevents testing the error dialog which should appear if you
try to delete a currency which is use (defined as a currency in a vendor
record). To test this error you could manually enter a currency code
into aqbooksellers.currency and then try to delete that currency.

Signed-off-by: Nicole C Engard <nengard@bywatersolutions.com>

Signed-off-by: Katrin Fischer <katrin.fischer.83@web.de>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>