Bug 16061: Use Font Awesome icons in reports when filter returns no results
In Reports -> Saved reports, when the filter form returns no results a
dialog is shown. This patch updates the dialog to use Font Awesome
icons.

Also changed: The dialog now includes a button for creating a new SQL
report, not just one to create a new guided report. The "clear" link in
the sidebar has been modified to pass HTML validation.

To test, apply the patch and go to Reports -> Saved reports.

- In the left-hand sidebar filter, submit terms which will return no
  results.
- Verify that the resulting dialog is correctly styled.
- Confirm that all three buttons in the dialog work correctly: New
  guided report, new SQL report, and Cancel.

Signed-off-by: Marc Véron <veron@veron.ch>
Signed-off-by: Tomas Cohen Arazi <tomascohen@unc.edu.ar>
I'm passing this one, while I don't like it is hardcoding a URL path,
because it will make is easier to spot it when time comes and we
make Koha delpoyable in any endpoint.

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com