Bug 16022 - Use Font Awesome icons on patron lists page
The patron lists page had one stray instance of a Glyphicon class
instead of Font Awesome. This patch corrects it.

To test, apply the patch and go to Tools -> Patron lists.

In the list of patron lists, confirm that the "Actions" menu displays
the correct icon for "Print patron cards" and that the function of the
menu item is correct.

Signed-off-by: Nicole C Engard <nengard@bywatersolutions.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>