Bug 15993 [Follow-up] Making currency actions buttons
This trivial follow-up adds an "actions" class to the table cell
containing the new buttons. In conjunction with the style update in Bug
16048, this will prevent the buttons from wrapping at narrow browser
window widths.

To test, apply the patch for Bug 16048 and go to Administration ->
Currencies and exchange rates.

In the table of currencies, confirm that the "Edit" and "Delete" buttons
are not stacked one on top of the other even at narrow browser window
widths.

Signed-off-by: Aleisha <aleishaamohia@hotmail.com>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com