Bug 15989 [Follow-up] Making classification sources actions buttons
This trivial follow-up adds an "actions" class to the table cell
containing the new buttons. In conjunction with the style update in Bug
16048, this will prevent the buttons from wrapping at narrow browser
window widths.

Also changed in this patch: A couple of capitalization errors.

To test, apply the patch for Bug 16048 and go to Administration ->
Classification sources. Confirm that the "Edit" and "Delete" buttons do
not stack on top of each other even at narrow browser widths.

Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>