Bug 16071: Use Font Awesome icons in authority duplicate warning dialog
In the authority editor, when a duplicate record is suspected a
dialog is shown. This patch updates the template to show Font Awesome
icons in this dialog.

Also changed: Using <h3> instead of <h4> in dialog so that it is
correctly styled. Adding "popup" class to duplicate record link to
indicate that a new window will be opened.

To test, apply the patch and go to Authorities -> New record.

- Use Z39.50 to import a record which is already in your catalog.
- Save the record. Confirm that the duplicate warning dialog is styled
  correctly.
- Confirm that both the "Save as new" and "Edit existing" buttons work
  correctly.

Signed-off-by: Hector Castro <hector.hecaxmmx@gmail.com>
Works as advertised

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>