Bug 16070: Default value for sysprefs should be an empty string
The default value for sysprefs should not be NULL but an empty string.
When a pref is cleared, it's set to an empty string, so it does not make
sense to create it with an undefined value.

The main purpose of this patch is to remove the warning in logs when a
pref is accessed for the first time and the cache is not yet populated.
It also ensures that the behavior will be the same for the first access
and the others.

Test plan:
SELECT COUNT(*) FROM systempreferences WHERE value IS NULL;
Should not return any results after the update DB entry executed.

Signed-off-by: Srdjan <srdjan@catalyst.net.nz>
Signed-off-by: Tomas Cohen Arazi <tomascohen@theke.io>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>