Bug 16078 [Revised] Remove unused YUI CSS
The YUI CSS file "skin.css" is no longer relevant to the staff client
templates and can be removed. Other references to YUI-related classes,
with the exception of grids, can also be removed.

This patch removes references in CSS to:

.yuimenuitemlabel
.yui-button
.yui-menu-button

...and removes the "yui-skin-sam" class from two templates' body tag.

skin.css is also included in a list of assets cached by the offline
circulation page. It can be safely removed.

To test, apply the patch and clear your cache if necessary. Look around
and confirm that the style of the staff client is unchanged.

Search for references to the listed classes and confirm that none are
found.

Revised patch squashes in the follow-up as well as a missing commit
removing skin.css from staff-global.css.

Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan Gallagher brendan@bywatersolutions.com