Bug 15793 - UX of circulation patron search with long lists of returned borrowers
Okay how about this:
This patch has no floating toolbar or names styled as links. I've removed the radio buttons. The row is clickable and will redirect you to the user's page.

EDIT: This patch makes the cursor a pointer when hovering over the patrons. The row will also be highlighted on hover to make it even clearer.
EDIT2: Removes unnecessary 'Select' button

To test:
1) Put in a patron search that will bring back a long list of patrons (ie the letter 'a')
2) Select a patron by clicking anywhere on the row - confirm you are taken to the correct patron page

Sponsored-by: Catalyst IT

Signed-off-by: Marc Véron <veron@veron.ch>
Signed-off-by: Liz Rea <liz@catalyst.net.nz>
Submit button removed, patron selection on this screen much better for long lists (and short ones as well.)

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>