Bug 16058: Delete individual news item from koha-news.pl
This patch adds a confirm message for deleting a news item from the home page in the staff client. It also gives the ability to delete an individual news item from koha-news.pl with a confirm message. The buttons on koha-news.pl are now styled as font awesome buttons.

To test:
1) Go to main intranet page. Click 'Delete' for a news item.
2) Confirm that Cancel stops the Delete, and OK deletes the item and takes you to tools/koha-news.pl
3) Confirm that the buttons (Edit and Delete) on koha-news.pl are now styled as font awesome buttons and the buttons don't wrap on a narrow browser.
4) Confirm that the column heading for the buttons is now "Actions".
5) Confirm that clicking Delete triggers a confirm message. Confirm that Cancel stops the Delete, and OK deletes the item and it no longer shows in the News table or anywhere else.

Sponsored-by: Catalyst IT

Signed-off-by: Marc Véron <veron@veron.ch>

Signed-off-by: Jonathan Druart <jonathan.druart@bugs.koha-community.org>
Amended-patch: just split the td on several lines for readability.

Signed-off-by: Brendan A Gallagher <brendan@bywatersolutions.com>